dlcs-docs-client/requirements.txt
python-dotenv~=1.2.1
requests